Cleared out Requests for hummingbird from anilist, started reimplementing
First off, cleared the request/response pairs from the anilist code to
start off fresh, and implemented the authorize and refresh code paths
needed.

I refactored a small bit of the Token input from twitter for easier use
with AniList as well.

I've also started laying down the needed code for the timer interaction
to make sure we're refreshing the anilist token every hour. Still need to
figure out how that works properly in Taiga so I can make sure I only
refresh it when it needs refreshed. Probably adding the timer when we get
the auth request done in the first place should we make it.